pyproject.toml
[tool.poetry]
name = "astar-monitor-django"
version = "0.1.0"
description = ""
authors = ["Nikita Ustiuzhanin <zaebumba@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
celery = "^5.3.4"
redis = "^5.0.0"
django-celery-beat = "^2.5.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
python-telegram-bot = "^13.15"
psycopg2 = "^2.9.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
